[NOM-17] - Big overhaul
