+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:nomad/screens/home_screen.dart';
+import 'package:nomad/screens/home/home_screen.dart';
 
 void main() {
   runApp(const App());
@@ -10,15 +10,8 @@ class App extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      theme: ThemeData.dark().copyWith(
-        appBarTheme: const AppBarTheme(
-          color: Color(0xFF141A3C),
-        ),
-        scaffoldBackgroundColor: const Color(0xFF141A3C),
-        useMaterial3: true,
-      ),
-      home: const HomeScreen(),
+    return const MaterialApp(
+      home: HomeScreen(),
     );
   }
 }
