+++ lib/screens/route_view/widgets/route_section_title.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class RouteSectionTitle extends StatelessWidget {
+  const RouteSectionTitle({super.key, required this.title});
+
+  final String title;
+
+  @override
+  Widget build(BuildContext context) {
+    return Align(
+      alignment: Alignment.topLeft,
+      child: Text(
+        title,
+        style: const TextStyle(fontSize: 16),
+      ),
+    );
+  }
+}
