+++ lib/screens/select_city/widgets/scrollable_bottom_sheet.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../constants.dart';
+import '../../../constants.dart';
 
 class ScrollableBottomSheet extends StatefulWidget {
   ScrollableBottomSheet({
